requirements.txt
streamlit==1.36.0
reportlab==4.2.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
networkx==3.3
ollama==0.3.0